refactor: Codebase combing and docs update
