+++ generals/agents/expander_agent.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from generals.core.config import Action, Direction
+from generals.core.game import Action, Direction
 from generals.core.observation import Observation
 
 from .agent import Agent
